Ejemplos MySql usando PDO

+++ Codigo Fuente/Unidad 14/base_datos.sql
@@ -1,5 +1,5 @@
 // Creacion de la tabla
-CREATE TABLE `peliculas`.`pelicula` ( `id` INT NOT NULL AUTO_INCREMENT , `titulo` VARCHAR(100) NOT NULL , `fecha_estreno` DATETIME NOT NULL , `calificacion` DECIMAL NOT NULL , `nacionalidad` VARCHAR(50) NOT NULL , `poster` VARCHAR(100) NOT NULL , PRIMARY KEY (`id`)) ENGINE = InnoDB;
+CREATE TABLE `peliculas`.`pelicula` ( `id` INT NOT NULL AUTO_INCREMENT , `titulo` VARCHAR(100) NOT NULL , `fecha_estreno` DATETIME NOT NULL , `calificacion` DECIMAL(2,1) NOT NULL , `nacionalidad` VARCHAR(50) NOT NULL , `poster` VARCHAR(100) NOT NULL , PRIMARY KEY (`id`)) ENGINE = InnoDB;
 
 // Insertamos dos registros
 INSERT INTO `pelicula`
